Create and use schema file for type
