+++ pages/entry/[entryId].tsx
@@ -8,6 +8,7 @@ import { remark } from 'remark';
 import html from 'remark-html';
 import GuideEntryBlock from '../../components/GuideEntryBlock';
 import testEntries from '../../data/test_entries.json';
+import {Author} from "../lib/Schema"
 
 export default function EntryPage({ entry }: InferGetStaticPropsType<typeof getStaticProps>) {
   return (
@@ -23,10 +24,6 @@ export default function EntryPage({ entry }: InferGetStaticPropsType<typeof getS
   );
 }
 
-type Author = {
-  name: string;
-  email: string;
-};
 
 type GuideEntry = {
   slug: string;
